requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
boto3==1.35.10
botocore==1.35.63
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.7
cron-descriptor==1.4.5
decorator==4.4.2
Django==5.1.3
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
drf-yasg==1.21.8
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
inflection==0.5.1
jmespath==1.0.1
kombu==5.4.2
moviepy==1.0.3
numpy==2.1.3
packaging==24.2
pillow==11.0.0
proglog==0.1.10
prompt_toolkit==3.0.48
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.3
setuptools==75.5.0
six==1.16.0
sqlparse==0.5.2
tqdm==4.67.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
